+++ src/main/kotlin/com/github/fmueller/jarvis/commands/ModelCommand.kt
@@ -7,12 +7,8 @@ import com.github.fmueller.jarvis.conversation.Message
 class ModelCommand(private val modelName: String) : SlashCommand {
 
     override suspend fun run(conversation: Conversation): Conversation {
-        if (modelName == "llama3.1" || modelName == "llama3.2") {
-            OllamaService.modelName = modelName
-            conversation.addMessage(Message.fromAssistant("Model changed to $modelName"))
-        } else {
-            conversation.addMessage(Message.fromAssistant("Invalid model name: $modelName"))
-        }
+        OllamaService.modelName = modelName
+        conversation.addMessage(Message.fromAssistant("Model changed to $modelName"))
         return conversation
     }
 }